pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-trace"
version = "0.1.0"
description = "Zero-configuration OpenTelemetry observability for Claude Agent SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "claude-trace contributors"}]
keywords = ["anthropic", "claude", "opentelemetry", "observability", "agents", "tracing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "anthropic>=0.40.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.9",
    "ruff>=0.4",
    "opentelemetry-exporter-otlp>=1.25.0",
    "rich>=13.0",
]
console = ["rich>=13.0"]
otlp = ["opentelemetry-exporter-otlp>=1.25.0"]

[project.urls]
Homepage = "https://github.com/claude-trace/claude-trace"
Documentation = "https://claude-trace.readthedocs.io"
Repository = "https://github.com/claude-trace/claude-trace"
Issues = "https://github.com/claude-trace/claude-trace/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_trace"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "PT"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
